Anasayfa tasarımı Default Index olsuturuldu fakat suan bir sıkıntı var donuyor acılmıyor.Dunku Amount problemi cozuldu.

+++ EntitityLayer/Entities/Car.cs
@@ -16,11 +16,8 @@ namespace EntitityLayer.Entities
         public int Km { get; set; }
         public string Transmission { get; set; }
         public byte Seat { get; set; }
-        public byte Luggage { get; set; }
+        public int Luggage { get; set; }
         public string Fuel { get; set; }
-       
-
-
         public string BigImageUrl { get; set; }
         public List<CarFeature> CarFeatures { get; set; }
         public List<CarDescription> CarDescriptions { get; set; }
